Bustage fix

+++ toolkit/components/passwordmgr/test/unit/test_storage_legacy_1.js
@@ -59,7 +59,8 @@ LoginTest.initStorage(storage, OUTDIR, filename);
 
 LoginTest.checkStorageData(storage, [], []);
 
-file.remove(false);
+if (file.exists())
+    file.remove(false);
 
 /* ========== 3 ========== */
 testnum++;
